Extract bin type
Found switch statement on bin type IDs in the Council's javascript. They do this to show what icons to show for each bin. Called the item URL to check what name corresponds with what ID.

In theory, we could look up ID <-> name mapping online, in case it changes. But the assignment between names and the types I'm using is a bit arbitrary so not something we can read off. It will make building a tidbyt app easier since we won't have to infer types there.